refactor(git-std): split cli/bump.rs into focused modules
